refactor(cryptography): one stateless program per curve
Cut the category down to raw syscall demos: bn254 (G2 add/mul plus
aggregate BLS verify, keeping the pairing op covered) and bls12-381
(G1/G2 group ops). The stateful application examples (multisig, key
registry, encrypted ballot) added account plumbing without teaching
anything new about the syscalls; they stay available in the upstream
crypto-primitives-examples repo, now linked from the category README.

Renames alt-bn128-g2 to bn254 to match the primitive naming.
