--- cryptography/bls-key-registry/pinocchio/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "types": ["mocha", "chai", "node"],
-        "typeRoots": ["./node_modules/@types"],
-        "lib": ["esnext"],
-        "module": "esnext",
-        "target": "esnext",
-        "moduleResolution": "bundler",
-        "esModuleInterop": true,
-        "resolveJsonModule": true,
-        "allowImportingTsExtensions": true,
-        "noEmit": true,
-        "skipLibCheck": true
-    }
-}
