--- cryptography/bls-multisig/pinocchio/program/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "bls-multisig-pinocchio-program"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-
-[features]
-no-entrypoint = []
-
-[dependencies]
-pinocchio.workspace = true
-solana-address.workspace = true
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = ['cfg(target_os, values("solana"))'] }
-
-[dev-dependencies]
-litesvm = "0.15"
-solana-account = "4.3"
-solana-instruction = "3.4"
-solana-keypair = "3.1"
-solana-native-token = "3.0"
-solana-signer = "3.0"
-solana-transaction = "4.1"
